chore: Refactor event date time form for create event
