+++ src/modules/organization/createEvent/event-date-time-form/EventDateTimeForm.vue
@@ -32,7 +32,7 @@
 </template>
 
 <script lang="ts">
-import { computed, defineComponent, Ref, ref, watch } from "vue";
+import { defineComponent, toRefs } from "vue";
 import FormOptionButton from "../form-option-button/FormOptionButton.vue";
 import FormSpecifyDateTime from "../form-specify-date-time/FormSpecifyDateTime.vue";
 import FormStartEndDateTime from "../form-start-end-date-time/FormStartEndDateTime.vue";
@@ -41,6 +41,7 @@ import {
   UPDATE_MODEL_VALUE
 } from "@/commons/constant";
 import { EventDurationsForm } from "@/commons/Interfaces";
+import useEventDateTimeForm from "./useEventDateTimeForm";
 
 export default defineComponent({
   name: "EventDateTimeForm",
@@ -57,75 +58,25 @@ export default defineComponent({
   },
   emits: [UPDATE_MODEL_VALUE],
   setup(props, context) {
-    const currentOption: Ref<EventDateTimeFormOption> = ref(
-      EventDateTimeFormOption.SPECIFY
-    );
+    const { modelValue } = toRefs(props);
 
-    const specifyEventDurations: Ref<EventDurationsForm> = ref({
-      durations: props.modelValue.durations,
-      isValid: props.modelValue.isValid
-    });
-
-    const startEndEventDurations: Ref<EventDurationsForm> = ref({
-      durations: [],
-      isValid: false
-    });
-
-    const announceLaterEventDuration: EventDurationsForm = {
-      durations: [],
-      isValid: true
-    };
-
-    const isSpecifyOption = computed(() => {
-      return currentOption.value === EventDateTimeFormOption.SPECIFY;
-    });
-
-    const isStartEndOption = computed(() => {
-      return currentOption.value === EventDateTimeFormOption.START_END;
-    });
-
-    const isLaterOption = computed(() => {
-      return currentOption.value === EventDateTimeFormOption.LATER;
-    });
-
-    function changeOption(option: EventDateTimeFormOption) {
-      currentOption.value = option;
-    }
-
-    watch(
-      () => specifyEventDurations,
-      () => {
-        if (isSpecifyOption.value)
-          context.emit(UPDATE_MODEL_VALUE, specifyEventDurations.value);
-      },
-      { deep: true }
-    );
-
-    watch(
-      () => startEndEventDurations,
-      () => {
-        if (isStartEndOption.value)
-          context.emit(UPDATE_MODEL_VALUE, startEndEventDurations.value);
-      },
-      { deep: true }
-    );
-
-    watch(currentOption, () => {
-      if (currentOption.value === EventDateTimeFormOption.SPECIFY)
-        context.emit(UPDATE_MODEL_VALUE, specifyEventDurations.value);
-      else if (currentOption.value === EventDateTimeFormOption.START_END)
-        context.emit(UPDATE_MODEL_VALUE, startEndEventDurations.value);
-      else context.emit(UPDATE_MODEL_VALUE, announceLaterEventDuration);
-    });
+    const {
+      isSpecifyOption,
+      isStartEndOption,
+      isLaterOption,
+      changeOption,
+      specifyEventDurations,
+      startEndEventDurations
+    } = useEventDateTimeForm(modelValue, context);
 
     return {
       isSpecifyOption,
       isStartEndOption,
       isLaterOption,
       changeOption,
-      EventDateTimeFormOption,
       specifyEventDurations,
-      startEndEventDurations
+      startEndEventDurations,
+      EventDateTimeFormOption
     };
   }
 });
