+++ src/modules/organization/createEvent/form-specify-date-time/FormSpecifyDateTime.vue
@@ -54,17 +54,15 @@
 </template>
 
 <script lang="ts">
-import { computed, defineComponent, ref, reactive, watch } from "vue";
+import { defineComponent, toRefs } from "vue";
 import BaseDatePicker from "@/commons/UI/BaseDatePicker.vue";
 import SingleNameSelect from "@/commons/UI/select/SingleNameSelect.vue";
 import BaseButton from "@/commons/UI/BaseButton.vue";
 import FormSpecifyDateTimeSchedule from "../form-specify-date-time-shcedule/FormSpecifyDateTimeSchedule.vue";
 import { hourNames, hourValues } from "@/commons/constant/hour";
-import { set, isEqual } from "date-fns";
-import { setTimeToZero } from "@/commons/utils/date";
-import { SetEventDurationsDurationInput } from "@/apollo/types";
 import { EventDurationsForm } from "@/commons/Interfaces";
 import { UPDATE_MODEL_VALUE } from "@/commons/constant";
+import useFormSpecifyDateTime from "./useFormSpecifyDateTime";
 
 export default defineComponent({
   name: "FormSpecifyDateTime",
@@ -82,72 +80,18 @@ export default defineComponent({
   },
   emits: [UPDATE_MODEL_VALUE],
   setup(props, context) {
-    const date = ref(setTimeToZero(new Date()).toString());
-    const startTime = ref(0);
-    const endTime = ref(0);
-    const durations: SetEventDurationsDurationInput[] = reactive(
-      props.modelValue.durations
-    );
+    const { modelValue } = toRefs(props);
 
-    const isValidTime = computed(() => {
-      return endTime.value > startTime.value;
-    });
-
-    const hasDurations = computed(() => {
-      return durations.length !== 0;
-    });
-
-    const inputDuration = computed(() => {
-      if (!isValidTime.value)
-        return {
-          start: new Date(),
-          finish: new Date()
-        };
-      const start = set(new Date(date.value), { hours: startTime.value });
-      const finish = set(new Date(date.value), { hours: endTime.value });
-      return { start: start, finish: finish };
-    });
-
-    function isEqualToInputDuration(
-      insertedDuration: SetEventDurationsDurationInput
-    ) {
-      const insertedDurationStartTime = new Date(insertedDuration.start);
-      const insertedDurationFinishTime = new Date(insertedDuration.finish);
-      return (
-        isEqual(inputDuration.value.start, insertedDurationStartTime) &&
-        isEqual(inputDuration.value.finish, insertedDurationFinishTime)
-      );
-    }
-
-    function addDuration() {
-      //Check if the duration is already inserted
-      const isDurationExist = durations.some(duration =>
-        isEqualToInputDuration(duration)
-      );
-      if (isDurationExist) return;
-      const inputDurationString = {
-        start: inputDuration.value.start.toString(),
-        finish: inputDuration.value.finish.toString()
-      };
-      durations.push(inputDurationString);
-    }
-
-    function removeDuration(index: number) {
-      durations.splice(index, 1);
-    }
-
-    const eventDurations: EventDurationsForm = reactive({
-      durations: durations,
-      isValid: hasDurations
-    });
-
-    watch(
-      () => eventDurations,
-      () => {
-        context.emit(UPDATE_MODEL_VALUE, eventDurations);
-      },
-      { deep: true }
-    );
+    const {
+      date,
+      startTime,
+      endTime,
+      isValidTime,
+      durations,
+      hasDurations,
+      addDuration,
+      removeDuration
+    } = useFormSpecifyDateTime(modelValue, context);
 
     return {
       date,
@@ -155,11 +99,11 @@ export default defineComponent({
       endTime,
       isValidTime,
       durations,
-      hourNames,
-      hourValues,
       addDuration,
       removeDuration,
-      hasDurations
+      hasDurations,
+      hourNames,
+      hourValues
     };
   }
 });
