+++ src/modules/organization/createEvent/form-start-end-date-time/FormStartEndDateTime.vue
@@ -44,15 +44,13 @@
 </template>
 
 <script lang="ts">
-import { computed, defineComponent, reactive, ref, watch } from "vue";
+import { defineComponent } from "vue";
 import BaseDatePicker from "@/commons/UI/BaseDatePicker.vue";
 import SingleNameSelect from "@/commons/UI/select/SingleNameSelect.vue";
 import { hourNames, hourValues } from "@/commons/constant/hour";
-import { isAfter, isEqual, set, differenceInCalendarDays, add } from "date-fns";
-import { setTimeToZero } from "@/commons/utils/date";
-import { SetEventDurationsDurationInput } from "@/apollo/types";
 import { EventDurationsForm } from "@/commons/Interfaces";
 import { UPDATE_MODEL_VALUE } from "@/commons/constant";
+import useFormStartEndDateTime from "./useFormStartEndDateTime";
 
 export default defineComponent({
   name: "FormStartEndDateTime",
@@ -68,65 +66,8 @@ export default defineComponent({
   },
   emits: [UPDATE_MODEL_VALUE],
   setup(_, context) {
-    const startDate = ref(setTimeToZero(new Date()).toString());
-    const endDate = ref(setTimeToZero(new Date()).toString());
-    const startTime = ref(0);
-    const endTime = ref(0);
-
-    const startFinishDate = computed(() => {
-      return {
-        start: new Date(startDate.value),
-        finish: new Date(endDate.value)
-      };
-    });
-
-    const isValidDuration = computed(() => {
-      const isValidDates =
-        isAfter(startFinishDate.value.finish, startFinishDate.value.start) ||
-        isEqual(startFinishDate.value.finish, startFinishDate.value.start);
-      const isValidTime = endTime.value > startTime.value;
-      return isValidDates && isValidTime;
-    });
-
-    function getStartFinishDurationOfADay(
-      day: Date,
-      startHour: number,
-      endHour: number
-    ): SetEventDurationsDurationInput {
-      return {
-        start: set(day, { hours: startHour }).toString(),
-        finish: set(day, { hours: endHour }).toString()
-      };
-    }
-
-    const durations = computed(() => {
-      const repeatedDurations: SetEventDurationsDurationInput[] = [];
-      if (!isValidDuration.value) return repeatedDurations;
-      const numOfDaysDifference = differenceInCalendarDays(
-        startFinishDate.value.finish,
-        startFinishDate.value.start
-      );
-      for (const dayFromStart of Array(numOfDaysDifference + 1).keys()) {
-        const day = add(startFinishDate.value.start, { days: dayFromStart });
-        const duration = getStartFinishDurationOfADay(
-          day,
-          startTime.value,
-          endTime.value
-        );
-        repeatedDurations.push(duration);
-      }
-      return repeatedDurations;
-    });
-
-    const eventDurations: EventDurationsForm = reactive({
-      durations: durations,
-      isValid: isValidDuration
-    });
-
-    watch(
-      () => eventDurations,
-      () => context.emit(UPDATE_MODEL_VALUE, eventDurations),
-      { deep: true }
+    const { startDate, endDate, startTime, endTime } = useFormStartEndDateTime(
+      context
     );
 
     return { startDate, endDate, startTime, endTime, hourNames, hourValues };
